celery_queue/requirements.txt
amqp==5.1.1
attrs==19.3.0
Automat==0.8.0
billiard==3.6.4.0
blinker==1.4
boto3==1.24.32
botocore==1.27.32
celery==5.2.7
certifi==2019.11.28
cffi==1.15.1
chardet==3.0.4
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cloud-init==22.2
colorama==0.4.3
command-not-found==0.3
commonmark==0.9.1
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
dbus-python==1.2.16
distro==1.4.0
distro-info===0.23ubuntu1
dnspython==2.2.1
docarray==0.13.32
entrypoints==0.3
grpcio==1.47.0
grpcio-health-checking==1.47.0
grpcio-reflection==1.47.0
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
img2vec-pytorch==1.0.1
importlib-metadata==1.5.0
incremental==16.10.1
jina==3.6.15
Jinja2==2.10.1
jmespath==1.0.1
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
kombu==5.2.4
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
loguru==0.6.0
MarkupSafe==1.1.0
more-itertools==4.2.0
netifaces==0.10.4
numpy==1.23.1
oauthlib==3.1.0
opencv-python==4.6.0.66
packaging==21.3
pexpect==4.6.0
pgzip==0.3.2
Pillow==9.2.0
pinecone-client==2.0.11
prometheus-client==0.14.1
prompt-toolkit==3.0.30
protobuf==3.20.1
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.21
pycurl==7.45.1
pygit2==1.10.1
Pygments==2.12.0
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
pymongo==4.2.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==3.0.9
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.0+ubuntu0.20.4.8
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-dotenv==0.20.0
pytz==2022.1
PyYAML==6.0
requests==2.22.0
requests-unixsocket==0.2.0
rich==12.5.1
ruamel.yaml==0.16.6
s3transfer==0.6.0
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
six==1.14.0
sos==4.3
ssh-import-id==5.10
systemd-python==234
tabulate==0.8.10
timm==0.6.7
torch==1.12.1
torchvision==0.13.1
towhee==0.8.0
tqdm==4.64.1
Twisted==18.9.0
typing-extensions==4.3.0
ubuntu-advantage-tools==27.10
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
uvloop==0.16.0
vine==5.0.0
wadllib==1.3.3
wcwidth==0.2.5
zipp==1.0.0
zope.interface==4.7.1
